+++ HISTORY.md
@@ -1,6 +1,8 @@
 # Snap<em>!</em> (BYOB) History
 
 ## Development
+### 2018-10-23
+* Objects: Don't stamp if the canvas is too small (and would throw an error)
 
 ### 2018-10-22
 * "Bigger Data" library: Added "analyze" function for fast frequency distributions